+++ site/src/main.rs
@@ -334,7 +334,6 @@ fn prim_class(prim: Primitive) -> &'static str {
         Primitive::Transpose => code_font!("monadic-function trans"),
         Primitive::Both => code_font!("monadic-modifier bi"),
         Primitive::All => code_font!("dyadic-modifier pan"),
-        prim if prim.is_ocean() => code_font!("ocean-function"),
         prim if prim.class() == PrimClass::Stack && prim.modifier_args().is_none() => {
             code_font!("stack-function")
         }